+++ addon/locales/ru/forms/fd-navigation.js
@@ -4,4 +4,6 @@ export default {
   search: 'Поиск',
   'root-folder': 'Корневая папка',
   cancel: 'Отменить',
+  'create-prototype-button-text': 'Создать представления, формы и приложение',
+  'create-prototype-error': 'Не удалось создать прототип приложения',
 };