+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+
 ## [0.8.0-beta.1] - 2019-08-16
 ### Added
 * Add primitives resizing on cad diagram.