+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 ### Added
 * Add text and background colors for primitives on cad diagram.
 
+### Fixed
+* Fix cad class attributes and methods parsing.
+
 ## [0.8.0-beta.1] - 2019-08-16
 ### Added
 * Add primitives resizing on cad diagram.